pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "weather_landscape"
version = "0.1.0"
requires-python = ">=3.7"
dependencies = [
  "pillow>=10.4.0",
  "suntime",
]

[tool.setuptools]
packages = ["weather_landscape"]

[tool.ruff]
line-length = 80
indent-width = 2
